[#1]feat: implement perform drop with coordinator method
drop시 데이터를 destination path로 전달하는 기능 구현
dragInteractionEnabled 구현
